translations: remove duplicated language file for Norwegian Bokmål
Right now, there are exists two files for the same language.
Because of that, it shows warning in Weblate. While looking at this
website: http://people.skolelinux.org/pere/blog/Spr_kkoder_for_POSIX_locale_i_Norge.html

It is mentioned that nb_NO should be used and nb should not be used.